requirements.txt
Flask==3.0.0
flask-cors==4.0.0
Pillow==10.1.0
numpy>=1.26.2,<2.3.0
face-recognition==1.3.0
opencv-python==4.12.0.88